float64

+++ boxes/pango_layout.go
@@ -4,7 +4,7 @@ import pr "github.com/benoitkugler/go-weasyprint/style/properties"
 
 type PangoLayout struct {
 	Text                 string
-	JustificationSpacing float32
+	JustificationSpacing float64
 	Context              interface{} // to prevent import cycle
 	Style                pr.Properties
 }
