+++ style/properties/utils.go
@@ -65,10 +65,10 @@ func (d Dimension) ToValue() Value {
 	return Value{Dimension: d}
 }
 
-func FToD(f float32) Dimension { return Dimension{Value: Float(f), Unit: Scalar} }
+func FToD(f float64) Dimension { return Dimension{Value: Float(f), Unit: Scalar} }
 func SToV(s string) Value      { return Value{String: s} }
-func FToV(f float32) Value     { return FToD(f).ToValue() }
-func (f Float) ToValue() Value { return FToV(float32(f)) }
+func FToV(f float64) Value     { return FToD(f).ToValue() }
+func (f Float) ToValue() Value { return FToV(float64(f)) }
 
 func (v Value) ToMaybeFloat() MaybeFloat {
 	if v.String == "auto" {
@@ -77,7 +77,7 @@ func (v Value) ToMaybeFloat() MaybeFloat {
 	return v.Value
 }
 
-func NewColor(r, g, b, a float32) Color {
+func NewColor(r, g, b, a float64) Color {
 	return Color{RGBA: parser.RGBA{R: r, G: g, B: b, A: a}, Type: parser.ColorRGBA}
 }
