+++ style/validation/validation_test.go
@@ -564,7 +564,7 @@ var (
 	red          = pr.NewColor(1, 0, 0, 1)
 	lime         = pr.NewColor(0, 1, 0, 1)
 	blue         = pr.NewColor(0, 0, 1, 1)
-	pi   float32 = math.Pi
+	pi   float64 = math.Pi
 )
 
 func checkGradientGeneric(t *testing.T, css string, expected pr.Image) {
